# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Let's Link API
  version: 1.0.0
  description: REST API for the Let's Link social planning platform.
servers:
  - url: http://localhost:4000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        username:
          type: string
        bio:
          type: string
        location:
          type: string
        interests:
          type: string
        phone:
          type: string
        profile_picture_url:
          type: string
          format: uri
        role:
          type: string
          enum: [user, admin]
        created_date:
          type: string
          format: date-time
        updated_date:
          type: string
          format: date-time
      required: [id, email, full_name, role, created_date, updated_date]
...
    Vote:
      type: object
      properties:
        id:
          type: string
        poll_id:
          type: string
        user_id:
          type: string
        selected_option:
          type: string
        created_date:
          type: string
          format: date-time
        updated_date:
          type: string
          format: date-time
      required: [id, poll_id, user_id, selected_option]
paths:
...
  /users:
    get:
      security:
        - bearerAuth: []
      summary: List users
      parameters:
        - in: query
          name: sort
          schema:
            type: string
          description: Sort expression (e.g. -created_date)
      responses:
        '200':
          description: Array of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /users/me:
    get:
      security:
        - bearerAuth: []
      summary: Get the current user
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      security:
        - bearerAuth: []
      summary: Update the current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
    get:
      security:
        - bearerAuth: []
      summary: Get a user by id
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      security:
        - bearerAuth: []
      summary: Update a user by id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      security:
        - bearerAuth: []
      summary: Delete a user
      responses:
        '204':
          description: User removed
...
  /votes:
    get:
      security:
        - bearerAuth: []
      summary: List votes accessible to the caller
      responses:
        '200':
          description: Vote list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vote'
    post:
      security:
        - bearerAuth: []
      summary: Create a vote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vote'
      responses:
        '201':
          description: Vote created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vote'
  /votes/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
    patch:
      security:
        - bearerAuth: []
      summary: Update a vote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated vote
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vote'
    delete:
      security:
        - bearerAuth: []
      summary: Delete a vote
      responses:
        '204':
          description: Vote deleted